chg: [stix2 import] Share the observable value walker with both parsers
Collecting the values an observable carries is not specific to External
content: an Internal Indicator merged into an Observed Data has to read
them too. Moving the walk to the base parser is the whole change - the
External caller reaches the same code through inheritance.
